COZMO-15672: Radio comms may silently drop packets (#303)
* Cherry-pick of https://github.com/anki/cozmo-one/pull/6469 from cozmo-one to victor

* COZMO-15672: Radio comms may silently drop packets

Cherry-pick https://github.com/anki/cozmo-one/pull/6469 from cozmo-one to victor

* COZMO-15672: Update victor components to check for (ssize_t < 0)

* COZMO-15672: Use const where possible per anki best practices

+++ animProcess/src/cozmoAnim/micDataProcessor.cpp
@@ -368,7 +368,7 @@ void MicDataProcessor::Update()
   static constexpr int kMaxReceiveBytes = 2000;
   char receiveArray[kMaxReceiveBytes];
   
-  const int bytesReceived = _udpServer->Recv(receiveArray, kMaxReceiveBytes);
+  const ssize_t bytesReceived = _udpServer->Recv(receiveArray, kMaxReceiveBytes);
   if (bytesReceived == 2)
   {
     if (receiveArray[0] != '\0' || receiveArray[1] != '\0')
